requirements.txt
matplotlib
openpyxl
scipy
pandas
requests
numpy
Pillow
beautifulsoup4
scikit_learn
xlrd
